--- packages/circle-demo-webapp/app/components/UsdcIcon/UsdcIcon.tsx
@@ -1,32 +1,0 @@
-interface UsdcIconProps {
-  /** Size of the icon in pixels */
-  size?: number;
-}
-
-export function UsdcIcon({ size = 32 }: UsdcIconProps) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width={size}
-      height={size}
-      viewBox="0 0 223.576 223.576"
-    >
-      <path
-        d="M111.788 0c61.74 0 111.788 50.05 111.788 111.788s-50.049 111.788-111.788 111.788S0 173.527 0 111.788 50.05 0 111.788 0Z"
-        style={{
-          fillRule: 'evenodd',
-          strokeWidth: 0,
-          fill: '#2775ca',
-        }}
-      />
-      <path
-        d="M137.33 33.073c-2.52-.805-4.581.691-4.581 3.338v6.509c0 1.774 1.337 3.794 3.003 4.404 26.757 9.8 45.904 35.519 45.904 65.629s-19.147 55.827-45.904 65.628c-1.827.67-3.003 2.459-3.003 4.405v6.508c0 2.646 2.06 4.143 4.582 3.338 33.813-10.804 58.298-42.482 58.298-79.88s-24.485-69.075-58.298-79.878Zm-46.502 3.338c0-2.647-2.061-4.143-4.582-3.338-33.814 10.804-58.299 42.482-58.299 79.88s24.485 69.075 58.299 79.879c2.52.805 4.582-.692 4.582-3.338v-6.508c0-1.775-1.337-3.795-3.004-4.405-26.756-9.8-45.903-35.519-45.903-65.628s19.147-55.828 45.903-65.629c1.667-.61 3.004-2.63 3.004-4.404v-6.51Zm24.454 20.648h-6.987a3.493 3.493 0 0 0-3.493 3.493v10.823C90.95 73.344 81.997 82.56 81.997 94.386c0 15.307 9.253 21.246 28.783 23.874 13.25 2.17 17.021 5.024 17.021 12.564 0 7.539-6.4 12.565-15.42 12.565-12.197 0-16.211-5.338-17.692-12.246-.35-1.637-1.752-2.833-3.426-2.833h-7.969c-2.007 0-3.576 1.808-3.231 3.785 2.029 11.629 9.497 20.16 24.739 22.228v11.03a3.494 3.494 0 0 0 3.493 3.494h6.987a3.494 3.494 0 0 0 3.493-3.494v-11.035c14.429-2.292 23.647-12.337 23.647-24.75 0-16.334-9.938-21.818-29.13-24.444-14.163-2.056-16.902-5.37-16.902-11.995 0-6.28 4.796-10.737 14.049-10.737 8.369 0 13.182 2.92 15.17 9.656.444 1.503 1.79 2.566 3.357 2.566h7.349c2.049 0 3.644-1.883 3.207-3.886-2.327-10.644-9.5-17.031-20.747-19.046v-11.13a3.493 3.493 0 0 0-3.493-3.493Z"
-        style={{
-          fillRule: 'evenodd',
-          strokeWidth: 0,
-          fill: '#fff',
-        }}
-      />
-    </svg>
-  );
-}
