--- packages/circle-demo-webapp/app/components/UsdcIcon/index.ts
@@ -1,1 +1,0 @@
-export * from './UsdcIcon';
