--- packages/circle-demo-webapp/app/components/WalletHeader/index.ts
@@ -1,1 +1,0 @@
-export * from './WalletHeader';
